% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Full_Vectors}
\alias{Full_Vectors}
\title{Word embeddings for expanded dictionaries}
\format{A data frame with 56 rows and 901 variables:
\describe{
  \item{Dictionary}{Name of the dictionary that the word embeddings are representing.}
  \item{fasttext}{variables starting in fasttext are the word embedding dimensions for Fasttext trained on 2 million word vectors trained with subword information on Common Crawl (https://fasttext.cc/docs/en/english-vectors.html)}
  \item{Glove}{variables starting in Glove are the word embedding dimensions for Glove trained on Common Crawl (840B tokens, 2.2M vocab, cased, 300d vectors; https://nlp.stanford.edu/projects/glove/) (https://fasttext.cc/docs/en/english-vectors.html)}
  \item{Word2vec}{variables starting in Word2vec are the word embedding dimensions for Word2vec trained Google News (https://code.google.com/archive/p/word2vec/)}
  ...
}}
\usage{
Full_Vectors
}
\description{
Word embeddings for expanded dictionaries
}
\keyword{datasets}
